requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
psycopg[binary]>=3.1
python-dotenv>=1.0
gunicorn>=23.0
whitenoise>=6.6

